# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kopi Tiang Alam API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Category {
  id: number;
  name: string;
  createdAt: string;
}

export interface CategoryInput {
  /** @minLength 1 */
  name: string;
}

export type ProductStatus = (typeof ProductStatus)[keyof typeof ProductStatus];

export const ProductStatus = {
  active: "active",
  inactive: "inactive",
} as const;

export interface Product {
  id: number;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  categoryName?: string | null;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  image?: string | null;
  stock?: number;
  status: ProductStatus;
  createdAt: string;
}

export type ProductInputStatus =
  (typeof ProductInputStatus)[keyof typeof ProductInputStatus];

export const ProductInputStatus = {
  active: "active",
  inactive: "inactive",
} as const;

export interface ProductInput {
  /** @nullable */
  categoryId?: number | null;
  /** @minLength 1 */
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  image?: string | null;
  stock?: number;
  status: ProductInputStatus;
}

export type ProductUpdateStatus =
  (typeof ProductUpdateStatus)[keyof typeof ProductUpdateStatus];

export const ProductUpdateStatus = {
  active: "active",
  inactive: "inactive",
} as const;

export interface ProductUpdate {
  /** @nullable */
  categoryId?: number | null;
  name?: string;
  /** @nullable */
  description?: string | null;
  price?: number;
  /** @nullable */
  image?: string | null;
  stock?: number;
  status?: ProductUpdateStatus;
}

export interface OrderItem {
  id: number;
  productId: number;
  productName: string;
  /** @nullable */
  productImage?: string | null;
  quantity: number;
  price: number;
  subtotal: number;
}

export interface OrderItemInput {
  productId: number;
  quantity: number;
}

export type OrderOrderType =
  (typeof OrderOrderType)[keyof typeof OrderOrderType];

export const OrderOrderType = {
  dine_in: "dine_in",
  takeaway: "takeaway",
} as const;

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  processing: "processing",
  preparing: "preparing",
  ready: "ready",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface Order {
  id: number;
  orderNumber: string;
  customerName: string;
  /** @nullable */
  tableNumber?: string | null;
  orderType: OrderOrderType;
  status: OrderStatus;
  /** @nullable */
  notes?: string | null;
  totalPrice: number;
  items?: OrderItem[];
  createdAt: string;
}

export type OrderInputOrderType =
  (typeof OrderInputOrderType)[keyof typeof OrderInputOrderType];

export const OrderInputOrderType = {
  dine_in: "dine_in",
  takeaway: "takeaway",
} as const;

export interface OrderInput {
  /** @minLength 1 */
  customerName: string;
  /** @nullable */
  tableNumber?: string | null;
  orderType: OrderInputOrderType;
  /** @nullable */
  notes?: string | null;
  items: OrderItemInput[];
}

export type OrderStatusUpdateStatus =
  (typeof OrderStatusUpdateStatus)[keyof typeof OrderStatusUpdateStatus];

export const OrderStatusUpdateStatus = {
  pending: "pending",
  processing: "processing",
  preparing: "preparing",
  ready: "ready",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface OrderStatusUpdate {
  status: OrderStatusUpdateStatus;
}

export type CoffeeTableStatus =
  (typeof CoffeeTableStatus)[keyof typeof CoffeeTableStatus];

export const CoffeeTableStatus = {
  available: "available",
  occupied: "occupied",
} as const;

export interface CoffeeTable {
  id: number;
  tableNumber: string;
  /** @nullable */
  qrCode?: string | null;
  status: CoffeeTableStatus;
  createdAt: string;
}

export type TableInputStatus =
  (typeof TableInputStatus)[keyof typeof TableInputStatus];

export const TableInputStatus = {
  available: "available",
  occupied: "occupied",
} as const;

export interface TableInput {
  /** @minLength 1 */
  tableNumber: string;
  status?: TableInputStatus;
}

export interface AdminLoginInput {
  username: string;
  password: string;
}

export interface AuthToken {
  token: string;
  username: string;
}

export interface OrderStatusCount {
  status: string;
  count: number;
}

export interface DashboardStats {
  totalOrders: number;
  activeOrders: number;
  dailySales: number;
  recentOrders: Order[];
  ordersByStatus: OrderStatusCount[];
}

export type ListProductsParams = {
  categoryId?: number;
  search?: string;
};

export type ListAdminOrdersParams = {
  status?: string;
  date?: string;
};