# lib/api-zod/src/generated/types/authToken.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kopi Tiang Alam API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AuthToken {
  token: string;
  username: string;
}