# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kopi Tiang Alam API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from "./orderItem";
import type { OrderOrderType } from "./orderOrderType";
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: number;
  orderNumber: string;
  customerName: string;
  /** @nullable */
  tableNumber?: string | null;
  orderType: OrderOrderType;
  status: OrderStatus;
  /** @nullable */
  notes?: string | null;
  totalPrice: number;
  items?: OrderItem[];
  createdAt: string;
}