# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kopi Tiang Alam API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProductStatus } from "./productStatus";

export interface Product {
  id: number;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  categoryName?: string | null;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  image?: string | null;
  stock?: number;
  status: ProductStatus;
  createdAt: string;
}