# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kopi Tiang Alam API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProductInputStatus } from "./productInputStatus";

export interface ProductInput {
  /** @nullable */
  categoryId?: number | null;
  /** @minLength 1 */
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  image?: string | null;
  stock?: number;
  status: ProductInputStatus;
}