# lib/api-zod/src/generated/types/productUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kopi Tiang Alam API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProductUpdateStatus } from "./productUpdateStatus";

export interface ProductUpdate {
  /** @nullable */
  categoryId?: number | null;
  name?: string;
  /** @nullable */
  description?: string | null;
  price?: number;
  /** @nullable */
  image?: string | null;
  stock?: number;
  status?: ProductUpdateStatus;
}