# lib/api-zod/src/generated/types/tableInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kopi Tiang Alam API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TableInputStatus } from "./tableInputStatus";

export interface TableInput {
  /** @minLength 1 */
  tableNumber: string;
  status?: TableInputStatus;
}